Improvement: Add top right to laser
