+++ src/app/widgets/CncLaserTopRight/styles.styl
@@ -1,0 +1,62 @@
+@import "../styles.styl"
+
+$z-index-float = 500
+$visualizer-padding = 10px
+
+.visualizer-top-right
+    position: absolute
+    top: $visualizer-padding
+    right: $visualizer-padding
+    z-index: $z-index-float
+    text-align: left
+
+.top-right-div11
+    background: url("/images/18x18/ic_back_normal_18x18.png") no-repeat center/18px 18px
+    width: 18px
+    height: 18px
+    display: inline-block
+    margin-top: 5px
+
+.top-right-div12
+    height: 18px
+    display: inline-block
+    vertical-align: middle
+    margin-top: -10px
+
+.top-right-div2
+    margin-top: 64px
+    margin-left: 50px
+    height: 104px
+    width: 134px
+    background: #FFFFFF
+    border-radius: 4px
+    border: 1px solid #C8C8C8
+
+    .title
+        margin-top: 16px
+        margin-left: 16px
+        height: 16px;
+        font-size: 14px;
+        font-family: Roboto-Medium, Roboto;
+        font-weight: 500;
+        color: #282828;
+        line-height: 16px;
+
+    .content
+        margin-top: 12px
+        margin-left: 16px
+        height: 14px;
+        font-size: 14px;
+        font-family: Roboto-Regular, Roboto;
+        font-weight: 400
+        color: #282828;
+        line-height: 16px
+
+        > span {
+            vertical-align: top
+            margin-left: 4px
+        }
+        > input {
+            width: 14px
+            height: 14px
+        }
